agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = "Agent Service for Communication System"
authors = ["joefox@gmail.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.13"
pika = "^1.3.2"
mistralai = "^1.6.0"
python-dotenv = "^1.0.0"
protobuf = "^5.26.1"
grpcio = "^1.71.0"
grpcio-tools = "^1.71.0"
shared-models = {path = "../shared_models", develop = true}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
